edge-ai/pyproject.toml
[project]
name = "simulator_ai"
version = "0.1.0"
requires-python = ">=3.10, <3.13"
dependencies = [
    "tensorflow-cpu==2.18.0",
    "scikit-learn==1.6.1",
    "joblib==1.4.2",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "matplotlib==3.10.0",
    "pyserial==3.5",
    "paho-mqtt==1.6.1"
]
